Made a WinForm sample
This is a WinForm sample that has a Gecko(Mozilla) Browser with the GE
plugin, and to the left is a text box for recognized gestures and above
is a placement to add the Kinect Toolkit viewer. You will need to
install the Gecko reference. I'm still trying to find a way to get these
browser apps to work in wpf... once I do, street view will be a breeze.